Final Commit
Changed the Data Model so that all saving/loading happens through
AppDelegate
